Robot interface class
This class interfaces with the robot, providing acces to sensors and
robot functions. A debug class that shows the sensor readings is also
avilable. I didn't make any changes to the structure of MainApp, just
made sure methods maching what the user inputted are called. Of course
for now they are just dummy methods.